services/ingest/pyproject.toml
[project]
name = "design-generator-ingest"
version = "0.1.0"
description = "Data ingestion pipeline: VLM processing, embedding, and vector DB indexing"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "Pillow>=11",
    "qdrant-client>=1.13,<1.15",
    "redis>=5.2",
    "httpx>=0.28",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.18",
    "sentence-transformers>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6",
    "ruff>=0.9",
    "pyright>=1.1",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "B", "A", "SIM"]
ignore = ["ANN101", "ANN102"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
